Refactor "return of investments view model" in to factory
